+++ googlesource/source.go
@@ -60,20 +60,6 @@ type Key struct {
 	Offset  string
 }
 
-// clientI provides function to create BigQuery Client
-type clientI interface {
-	Client() (*bigquery.Client, error)
-}
-
-type client struct {
-	ctx       context.Context
-	projectID string
-	opts      []option.ClientOption
-}
-
-func (client *client) Client() (*bigquery.Client, error) {
-	return bigquery.NewClient(client.ctx, client.projectID, client.opts...)
-}
 func NewSource() sdk.Source {
 	return &Source{}
 }
@@ -144,7 +130,6 @@ func (s *Source) Ack(ctx context.Context, position sdk.Position) error {
 }
 
 func (s *Source) Teardown(ctx context.Context) error {
-
 	if s.records != nil {
 		close(s.records)
 	}
